fix the sub-d_sw unit tests
